test reporting period module
