+++ src/reporting-period/reporting-period.service.ts
@@ -1,4 +1,5 @@
 import { Injectable } from '@nestjs/common';
+import { LessThanOrEqual } from 'typeorm';
 import { ReportingPeriodDTO } from '../dto/reporting-period.dto';
 import { ReportingPeriodRepository } from './reporting-period.repository';
 
@@ -7,6 +8,13 @@ export class ReportingPeriodService {
   constructor(private readonly repository: ReportingPeriodRepository) {}
 
   async getReportingPeriods(): Promise<ReportingPeriodDTO[]> {
-    return await this.repository.getReportingPeriods();
+    const today = new Date(Date.now());
+    const currentYear = today.getFullYear();
+    const currentQuarter = Math.floor(today.getMonth() / 3 + 1);
+    const periodAbbreviation = `${currentYear} Q${currentQuarter}`;
+
+    return await this.repository.find({
+      where: { periodAbbreviation: LessThanOrEqual(periodAbbreviation) },
+    });
   }
 }
